Add simple Lua scripting support, and show containers in listing table
Lua scripting: Provider scripts can now be added to share/lua/provider.
Currently, they are used to provide functionality for the new "Product
Page" button that can be clicked to open a part's product page in an
external web browser. This only works if the product ID is found in a
special format in the 'Notes' property. See the Lua scripts for details.
Current providers: digikey, reichelt, eckstein-shop